Remove API views Modified urls, views and serializers py's Create API folder and transfer urls, views and serializer py's
